requirements.txt
asgiref==3.2.7
Django==3.0.6
djangorestframework==3.11.0
fonttools==4.11.0
fpdf==1.7.2
gunicorn==20.0.4
numpy==1.18.4
opencv-python==4.2.0.34
Pillow==7.1.2
PyPDF2==1.26.0
pytz==2020.1
sqlparse==0.3.1
